+++ src/tools/index.ts
@@ -52,6 +52,7 @@ import { registerCheckWritableCoords } from "./celldata/check-writable-coords.js
 import { registerCreateElement } from "./dimension-management/create-element.js";
 import { registerUpdateElement } from "./dimension-management/update-element.js";
 import { registerDeleteElement } from "./dimension-management/delete-element.js";
+import { registerDeleteElements } from "./dimension-management/delete-elements.js";
 import { registerMoveElement } from "./dimension-management/move-element.js";
 import { registerCreateDimension } from "./dimension-management/create-dimension.js";
 import { registerDeleteDimension } from "./dimension-management/delete-dimension.js";
@@ -198,6 +199,7 @@ const REGISTRARS: ToolRegistrar[] = [
   registerCreateElement,
   registerUpdateElement,
   registerDeleteElement,
+  registerDeleteElements,
   registerMoveElement,
   registerListElementAttributes,
   registerCreateElementAttribute,